refactor getSummaries
